refactor: rename HelperUtility to IconService
Move HelperUtility to Service/IconService with proper namespace, reflecting
its role as the central icon management service. Update all consumers and
DI configuration.
